loadcaptcha bugfix coa
git-svn-id: svn://svn.jdownloader.org/jdownloader/trunk@7454 ebf7c1c2-ba36-0410-9fe8-c592906822b4

+++ src/jd/captcha/easy/LoadCaptchas.java
@@ -441,10 +441,9 @@ public class LoadCaptchas {
             final ArrayList<JButton> bts = new ArrayList<JButton>();
             for (int j = 0; j < images.size(); j++) {
                 final LoadImage f = images.get(j);
-                if (!f.file.exists() || f.file.length() < 100) continue;
+                if (f == null || f.file == null || !f.file.exists() || f.file.length() < 100) continue;
                 final BufferedImage captchaImage = JDImage.getImage(f.file);
-                if(captchaImage==null)
-                {
+                if (captchaImage == null) {
                     f.file.delete();
                     continue;
                 }
